Implement skipping back over content-less branches.
This avoids the problem that an IO-dependent slide can
never be reverted (which was because reverting would just go
back to the content-devoid `Interactive` node, but evaluation that
would trigger an automatic progress, voiding the revert.

Now, we revert the entire superbranch, back to the last slide with
actual content in it.