Explicitly specify relative approot.
Yesod-1.5 has changed the default to `guessApproot`, which causes
a normal standalone program to spew out `localhost:` urls. That's ok
while running the server on the same machine as the viewer, but no
good if both are seperate. In particular, the demo presentation relies
on relative urls, because it runs yeamer behind a reverse proxy.

Discussed in https://github.com/yesodweb/yesod/issues/1549.