Simplify the type of `verbatim`.
Previously it could be `IPresentation m r` for any `Monoid r`, but
this appears to have not that much use and just necessitates awkward
type restrictions when used in `do` blocks, lest we get ambiguous
type variables.

Now it's just `IPresentation m ()`, which will be sufficient for pretty
much all needs.