Hard-bake some styling for generic-data display into Yeamer lib.
This is generally something we try to avoid, but the borders between record fields
require knowledge about the specific CSS-Grid architecture, which is something
that may well change in the future (tables?) so it would be bad to require users
to refer to this kind of implementation detail in their styles.

Also, the gentle dashed boundaries are probably quite uncontroversial.